tout est ok, reste les codes de reponse a mettre en place

+++ app/Http/Controllers/CategoryController.php
@@ -1,18 +1,18 @@
 <?php
 
 namespace App\Http\Controllers;
-
-
-
-
 use App\Models\Category;
 use Illuminate\Http\Request;
 
 class CategoryController extends CoreController
 {
 
-
-
+    /**
+     * create
+     *
+     * @param  mixed $request
+     * @return void
+     */
     public function create(Request $request){
 
         $categories = Category::create($request->all());
@@ -20,28 +20,88 @@ class CategoryController extends CoreController
         return response()->json($categories);
 
     }
-    
+    /**
+     * edit
+     *
+     * @param  mixed $id
+     * @return void
+     */
+    public function edit($id)
+    {
+        $category = Category::find($id);
+        if ($category !== null)
+        {
+            return response()->json($category);
+        }else {
+            abort(404,"Pas le bon ID !");
+        }
+
+    }
+    /**
+     * update
+     *
+     * @param  mixed $request
+     * @param  mixed $id
+     * @return void
+     */
     public function update(Request $request, $id){
         $category  = Category::find($id);
-        \dump($category);
-        $category->name = $request->input('name');
-        $category->status = $request->input('status');
-        \dd($category->name);
-
-        $category->save();
-
-        return response()->json($category);
+        if ($category !== null)
+        {
+            $category->name = $request->input('name');
+            $category->status = $request->input('status');
+            
+            $category->save();
+            return response()->json($category);
+        }else {
+            abort(404,"Pas le bon ID !");
+        }
     }
+
+    /**
+     * patch
+     *
+     * @param  mixed $request
+     * @param  mixed $id
+     * @return void
+     */
+    public function patch(Request $request, $id){
+        $category  = Category::find($id);
+        if ($category !== null)
+        {
+            $category->status = $request->input('status');
+            $category->save();
+            return response()->json($category);
+        }else {
+            abort(404,"Pas le bon ID !");
+        }
+
+    }
+    /**
+     * delete
+     *
+     * @param  mixed $id
+     * @return void
+     */
     public function delete($id){
         $category  = Category::find($id);
-        $category->delete();
+        if ($category !== null)
+        {
+            $category->delete();
+            return response()->json('Removed successfully.');
+        }else {
+            abort(404,"Pas le bon ID !");
+        }
 
-        return response()->json('Removed successfully.');
     }
+
+    /**
+     * list
+     *
+     * @return void
+     */
     public function list(){
-
         $categories  = Category::all();
-
         return response()->json($categories);
 
     }
